# runtime: one lock per order